Harden parser types, loops, and predicates (#55)
- Widen Marker/Event positions to usize, removing casts
- Widen n_raw_tokens to u32, removing truncation
- Add checkpoint-based progress guards to 6 loops
- Centralize type keyword and direction predicates
- Document fuel pattern and Ident heuristic

+++ crates/lyra-parser/src/lib.rs
@@ -88,7 +88,7 @@ fn replay(events: &[Event], tokens: &[Token], src: &str) -> rowan::GreenNode {
                 let mut idx = i;
                 let mut fp = *forward_parent;
                 while let Some(delta) = fp {
-                    idx += delta as usize;
+                    idx += delta;
                     processed[idx] = true;
                     if let Event::Start {
                         kind,
